% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pkgconfig}
\alias{pkgconfig}
\title{Compiler arguments for using Rzstdlib}
\usage{
pkgconfig(opt = c("PKG_C_LIBS", "PKG_RPATH"))
}
\arguments{
\item{opt}{A scalar character from the list of available options;
default is \code{PKG_C_LIBS}.}
}
\value{
\code{NULL}; prints the corresponding value to stdout.
}
\description{
This function returns values for
\code{PKG_CC_FLAGS} variable for use in Makevars files.
}
\examples{
pkgconfig("PKG_C_LIBS")
}
