% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstd.R
\name{swap_suffix}
\alias{swap_suffix}
\title{Silly little function that adds a suffix to a path if it's already there, and removes it if isn't}
\usage{
swap_suffix(file, suffix = "zstd")
}
\arguments{
\item{suffix}{}
}
\description{
Silly little function that adds a suffix to a path if it's already there, and removes it if isn't
}
