% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zstd.R
\name{zstd}
\alias{zstd}
\title{Compress/decompress a file using the zstd protocol}
\usage{
zstd(file, dest = NULL, level = 3L, overwrite = FALSE,
  dictionary = NULL, remove_source = FALSE, decompression = NA)
}
\arguments{
\item{file}{A character(1) path to an existing file. This file will be compressed using the ZSTD algorithm if it is uncompressed, and it will be decompressed if it is already compressed}

\item{dest}{A character(1) path to the destination file. If dest exists, then it is only over-written when overwrite=TRUE.}

\item{level}{an integer(1) argument in the closed set 1-19 specifying compression level. A higher compression level may result in a smaller file size, but will take longer to compute.  This is ignored in the case of decompression}

\item{overwrite}{a logical(1) indicating whether `dest` should be over-written if it already exists.}

\item{dictionary}{If specified, a character(1) path to an existing dictionary for dictionary-based compression.}

\item{remove_source}{A logical(1) indicating whether to remove `file` after compression/decompression is successful}

\item{decompression}{a logical(1) indicating whether to compress or decompress the file.  By default, this will be inferred based on the file extension}
}
\value{
The full path to `dest`
}
\description{
Compress/decompress a file using the zstd protocol
}
\examples{

# get some text
txt <- readLines(file.path(R.home("doc"), "COPYING"))
#generate a temporary file
tmpf <- fs::file_temp()
#write the text to the file
write(txt,tmpf)
#compress the result
dest <- zstd(tmpf)
}
